chore: split parser phases
